Separated lexicon from normalizers
